Switch to OpenSearch: updating references and imports ES -> OS
