+++ src/main/java/org/craftercms/engine/search/SiteAwareOpenSearchClient.java
@@ -13,57 +13,46 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+package org.craftercms.engine.search;
 
-package org.craftercms.engine.search.legacy;
-
-import org.apache.commons.lang3.ArrayUtils;
+import org.apache.commons.collections4.list.SetUniqueList;
 import org.apache.commons.lang3.StringUtils;
 import org.craftercms.commons.http.RequestContext;
 import org.craftercms.engine.service.context.SiteContext;
 import org.craftercms.engine.util.LocaleUtils;
-import org.craftercms.search.elasticsearch.impl.AbstractElasticsearchWrapper;
-import org.elasticsearch.action.search.SearchRequest;
-import org.elasticsearch.action.search.SearchType;
-import org.elasticsearch.action.support.IndicesOptions;
-import org.elasticsearch.client.RestHighLevelClient;
-import org.elasticsearch.index.query.BoolQueryBuilder;
-import org.elasticsearch.search.builder.SearchSourceBuilder;
+import org.craftercms.search.opensearch.impl.client.AbstractOpenSearchClientWrapper;
+import org.opensearch.client.opensearch.OpenSearchClient;
+import org.opensearch.client.opensearch._types.SearchType;
+import org.opensearch.client.opensearch._types.query_dsl.BoolQuery;
+import org.opensearch.client.opensearch.core.SearchRequest;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 
+import javax.servlet.http.HttpServletRequest;
 import java.beans.ConstructorProperties;
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Locale;
+import java.util.*;
 import java.util.stream.Stream;
 
 import static java.util.Collections.emptyList;
 import static java.util.stream.Collectors.joining;
+import static java.util.stream.Collectors.toMap;
 import static org.apache.commons.collections4.CollectionUtils.isNotEmpty;
 import static org.apache.commons.collections4.list.SetUniqueList.setUniqueList;
-import static org.apache.commons.lang3.StringUtils.appendIfMissing;
-import static org.apache.commons.lang3.StringUtils.removeStart;
-import static org.apache.commons.lang3.StringUtils.startsWith;
+import static org.apache.commons.lang3.StringUtils.*;
 import static org.craftercms.commons.locale.LocaleUtils.appendLocale;
 import static org.craftercms.commons.locale.LocaleUtils.getCompatibleLocales;
-import static org.craftercms.engine.util.LocaleUtils.getCurrentLocale;
-import static org.craftercms.engine.util.LocaleUtils.getDefaultLocale;
-import static org.craftercms.engine.util.LocaleUtils.isTranslationEnabled;
-import static org.elasticsearch.index.query.QueryBuilders.boolQuery;
-import static org.elasticsearch.index.query.QueryBuilders.existsQuery;
-import static org.elasticsearch.index.query.QueryBuilders.matchQuery;
+import static org.craftercms.engine.util.LocaleUtils.*;
 
 /**
- * Implementation of {@link org.craftercms.search.elasticsearch.ElasticsearchWrapper}
- * that sets the index based on the current site context for all search requests.
+ * Implementation of {@link AbstractOpenSearchClientWrapper} that sets the index and security filters based on the
+ * current site context for all search requests.
+ *
  * @author joseross
- * @since 3.1
+ * @since 4.0.0
  */
-public class SiteAwareElasticsearchService extends AbstractElasticsearchWrapper {
+public class SiteAwareOpenSearchClient extends AbstractOpenSearchClientWrapper {
 
     private static final String DEFAULT_ROLE_FIELD_NAME = "authorizedRoles.item.role";
 
@@ -72,7 +61,6 @@ public class SiteAwareElasticsearchService extends AbstractElasticsearchWrapper 
     private static final String DEFAULT_FALLBACK_PARAM_NAME = "localeFallback";
 
     private static final String ROLE_PREFIX = "ROLE_";
-
 
     /**
      * Format used to build the index id
@@ -88,7 +76,7 @@ public class SiteAwareElasticsearchService extends AbstractElasticsearchWrapper 
     protected final boolean enableTranslation;
 
     @ConstructorProperties({"client", "indexIdFormat", "enableTranslation"})
-    public SiteAwareElasticsearchService(RestHighLevelClient client, String indexIdFormat, boolean enableTranslation) {
+    public SiteAwareOpenSearchClient(OpenSearchClient client, String indexIdFormat, boolean enableTranslation) {
         super(client);
         this.indexIdFormat = indexIdFormat;
         this.enableTranslation = enableTranslation;
@@ -110,12 +98,12 @@ public class SiteAwareElasticsearchService extends AbstractElasticsearchWrapper 
         if (!(enableTranslation && isTranslationEnabled())) {
             return emptyList();
         }
-        var locales = setUniqueList(new LinkedList<Locale>());
-        var requestContext = RequestContext.getCurrent();
+        SetUniqueList<Locale> locales = setUniqueList(new LinkedList<>());
+        RequestContext requestContext = RequestContext.getCurrent();
         String useFallback = null;
         if (requestContext != null) {
-            var httpRequest = requestContext.getRequest();
-            var requestedLocales = httpRequest.getParameter(localesParameterName);
+            HttpServletRequest httpRequest = requestContext.getRequest();
+            String requestedLocales = httpRequest.getParameter(localesParameterName);
             useFallback = httpRequest.getParameter(fallbackParameterName);
             if (StringUtils.isNotEmpty(requestedLocales)) {
                 // split the locales and add all compatible versions to the list
@@ -136,11 +124,15 @@ public class SiteAwareElasticsearchService extends AbstractElasticsearchWrapper 
         return locales;
     }
 
-    /**
-     * {@inheritDoc}
-     */
+    protected String addPrefix(SiteContext siteContext, String name) {
+        return String.format("%s_%s", siteContext.getSiteName(), name);
+    }
+
     @Override
-    protected void updateIndex(final SearchRequest request) {
+    protected void updateIndex(SearchRequest request, Map<String, Object> parameters, RequestUpdates updates) {
+        // Call the parent to support custom index search
+        super.updateIndex(request, parameters, updates);
+
         SiteContext siteContext = SiteContext.getCurrent();
         if (siteContext == null) {
             throw new IllegalStateException("Current site context not found");
@@ -149,7 +141,7 @@ public class SiteAwareElasticsearchService extends AbstractElasticsearchWrapper 
         // Generate the default alias for the current site
         String aliasName = String.format(indexIdFormat, siteContext.getSiteName());
         // Get the requested indices
-        String[] currentIndices = request.indices();
+        List<String> currentIndices = updates.getIndex();
 
         // list of aliases to query
         var aliases = new LinkedList<String>();
@@ -164,80 +156,101 @@ public class SiteAwareElasticsearchService extends AbstractElasticsearchWrapper 
         // the original alias will always be included for backward compatibility
         aliases.add(aliasName);
 
-        if (ArrayUtils.isNotEmpty(currentIndices)) {
+        List<Map<String, Double>> boosting = new LinkedList<>();
+
+        if (isNotEmpty(currentIndices)) {
             // Add the site name prefix for all indices
-            Stream.of(currentIndices).map(index -> addPrefix(siteContext, index)).forEach(aliases::add);
+            currentIndices.stream().map(index -> addPrefix(siteContext, index)).forEach(aliases::add);
 
             // Add the site name prefix for the boosting if needed
-            List<SearchSourceBuilder.IndexBoost> indexBoosts = new ArrayList<>(request.source().indexBoosts());
-            if (isNotEmpty(indexBoosts)) {
-                indexBoosts.forEach(indexBoost ->
-                        request.source().indexBoost(addPrefix(siteContext, indexBoost.getIndex()),
-                                indexBoost.getBoost()));
+            List<Map<String, Double>> existingBoosting = request.indicesBoost();
+            if (isNotEmpty(existingBoosting)) {
+                existingBoosting.stream()
+                                .map(boost -> boost.entrySet().stream()
+                                        .collect(toMap(entry -> addPrefix(siteContext, entry.getKey()),
+                                                       Map.Entry::getValue)))
+                                .forEach(boosting::add);
             }
         }
 
         logger.debug("Executing query for aliases: {}", aliases);
 
         // Override the indices field in the request
-        request.indices(aliases.toArray(new String[0]));
+        updates.setIndex(aliases);
 
         if (aliases.size() > 1) {
             // Boost the results based on the index
-            var boost = 1f;
+            var boost = 1d;
             var iterator = aliases.listIterator(aliases.size());
             while (iterator.hasPrevious()) {
-                request.source().indexBoost(iterator.previous(), boost);
+                boosting.add(Map.of(iterator.previous(), boost));
                 // TODO: Make this value configurable per site
                 boost += 0.05;
             }
+            updates.setIndicesBoost(boosting);
 
             // Don't fail if one of the indices doesn't exist
-            IndicesOptions originalOptions = request.indicesOptions();
-            request.indicesOptions(IndicesOptions.fromOptions(true, originalOptions.allowNoIndices(),
-                    originalOptions.expandWildcardsOpen(), originalOptions.expandWildcardsClosed(),
-                    originalOptions.allowAliasesToMultipleIndices(), originalOptions.forbidClosedIndices(),
-                    originalOptions.ignoreAliases(), originalOptions.ignoreThrottled()));
+            updates.setIgnoreUnavailable(true);
 
             // Fix scores across multiple indices
-            request.searchType(SearchType.DFS_QUERY_THEN_FETCH);
+            updates.setSearchType(SearchType.DfsQueryThenFetch);
         }
-    }
-
-    protected String addPrefix(SiteContext siteContext, String name) {
-        return String.format("%s_%s", siteContext.getSiteName(), name);
     }
 
     @Override
-    protected void updateFilters(final SearchRequest request) {
-        super.updateFilters(request);
+    protected void updateQuery(SearchRequest request, Map<String, Object> parameters, RequestUpdates updates) {
+        super.updateQuery(request, parameters, updates);
 
-        BoolQueryBuilder mainQuery = (BoolQueryBuilder) request.source().query();
+        // Use the updated if it exists
+        BoolQuery mainQuery = Optional.ofNullable(updates.getQuery()).orElse(request.query()).bool();
 
-        Authentication auth = null;
-        SecurityContext context = SecurityContextHolder.getContext();
-        if (context != null) {
-            auth = context.getAuthentication();
-        }
+        Authentication auth = SecurityContextHolder.getContext() != null?
+                                SecurityContextHolder.getContext().getAuthentication() : null;
 
         // Include all public items
-        BoolQueryBuilder securityQuery = boolQuery()
-                .should(boolQuery().mustNot(existsQuery(roleFieldName)))
-                .should(matchQuery(roleFieldName, "anonymous"));
+        BoolQuery.Builder securityQuery = new BoolQuery.Builder()
+            .should(s -> s
+                .bool(b -> b
+                    .mustNot(n -> n
+                        .exists(e -> e
+                            .field(roleFieldName)
+                        )
+                    )
+                )
+            )
+            .should(s -> s
+                .match(m -> m
+                    .field(roleFieldName)
+                    .query(q -> q
+                        .stringValue("anonymous")
+                    )
+                )
+            );
 
         if (auth != null && !(auth instanceof AnonymousAuthenticationToken) && isNotEmpty(auth.getAuthorities())) {
             logger.debug("Filtering search results for roles: {}", auth.getAuthorities());
-            securityQuery.should(matchQuery(roleFieldName, auth.getAuthorities().stream()
+            securityQuery.should(s -> s
+                .match(m -> m
+                    .field(roleFieldName)
+                    .query(q -> q
+                        .stringValue(auth.getAuthorities().stream()
                             .map(GrantedAuthority::getAuthority)
-                            .map(role -> role +  " " +
-                                    (startsWith(role, ROLE_PREFIX)? removeStart(role, ROLE_PREFIX)
-                                            : appendIfMissing(role, ROLE_PREFIX)))
-                            .collect(joining(" "))));
+                            .map(role -> role +  StringUtils.SPACE +
+                                (startsWith(role, ROLE_PREFIX)? removeStart(role, ROLE_PREFIX)
+                                                                : appendIfMissing(role, ROLE_PREFIX)))
+                            .collect(joining(StringUtils.SPACE)))
+                    )
+                )
+            );
         } else {
             logger.debug("Filtering search to show only public items");
         }
-
-        mainQuery.filter(boolQuery().must(securityQuery));
+        updates.setQuery(q -> q
+            .bool(b -> b
+                .must(mainQuery._toQuery())
+                .filter(securityQuery.build()._toQuery())
+            )
+        );
     }
 
 }
