+++ src/main/resources/crafter/engine/rendering/controller-context.xml
@@ -117,7 +117,7 @@
 
     <bean id="crafter.siteElasticsearchController"
           class="org.craftercms.engine.controller.rest.SiteElasticsearchController">
-        <property name="elasticsearchService" ref="crafter.elasticsearchService"/>
+        <property name="openSearchService" ref="crafter.elasticsearchService"/>
     </bean>
 
     <bean id="crafter.siteGraphQLController" class="org.craftercms.engine.controller.rest.SiteGraphQLController"/>
